Fix #to_param for invalid models
#to_param returns the slug if the model is valid.

If the model is invalid, #to_param returns the previous slug value.
When an update fails, the new slug value is not yet persisted. In that
case, you want the previous #to_param/slug value, so that paths/urls
for the model are valid.

Before, when #to_param was called in an after_* callback, calling
#slug_was caused a warning.